Removed for some time migration + added publishState column in FullEventEntity

+++ app/src/main/java/com/pawlowski/stuboard/presentation/use_cases/GetEventPublishingStatusUseCase.kt
@@ -3,4 +3,4 @@ package com.pawlowski.stuboard.presentation.use_cases
 import com.pawlowski.stuboard.presentation.my_events.EventPublishState
 import kotlinx.coroutines.flow.Flow
 
-fun interface GetEventPublishingStatusUseCase: () -> Flow<EventPublishState>
+fun interface GetEventPublishingStatusUseCase: (Int) -> Flow<EventPublishState>
